# perf: optimise `noImportCycles`