--- a/.changeset/cyclic-checks-charge.md
+++ b/.changeset/cyclic-checks-charge.md
@@ -1,0 +1,5 @@
+---
+"@biomejs/biome": patch
+---
+
+Improved the performance of `noImportCycles` by ~30%.