… Cargo.lock
… crates/biome_module_graph/tests/snapshots/test_export_const_type_declaration_with_namespace.snap
… crates/biome_module_graph/tests/snapshots/test_resolve_export_types.snap
… crates/biome_module_graph/tests/snapshots/test_resolve_exports.snap
… crates/biome_module_graph/tests/snapshots/test_resolve_merged_types.snap
… crates/biome_module_graph/tests/snapshots/test_resolve_multiple_reexports.snap
… crates/biome_module_graph/tests/snapshots/test_resolve_recursive_looking_country_info.snap
… crates/biome_module_graph/tests/snapshots/test_resolve_recursive_looking_vfile.snap