## Summary

Improved the performance of `noImportCycles` by ~30%.

The majority of the savings are achieved by using an `IndexMap` instead of `BTreeMap` for storing the imports in the module info. These savings are likely to have some positive impact on type inference as well, although I expect the savings to be less pronounced there.

The rest of the savings were achieved by using an `FxHashSet` instead of a regular `HashSet` to track the paths that have been already been seen during cycle detection. Also, I'm storing the `resolved_path` in the set now, since it has an internal `Arc`, meaning we don't need another allocation for storing it, and we can call `.insert()` directly without needing to call `.contains()` first, since that was also just an attempt at avoiding the allocation.

Note that the performance of `noImportCycles` appears to fluctuate _highly_. On `main`, with a largish repository, I got timings between 12s and 28s. With these fixes it appears to be between 8s and 18s. Still widely inconsistent, but it appears to be a clear improvement :)

## Test Plan

Everything should stay green.